all: fix implicit signedness conversions
Clang diagnoses signedness changes in implicit conversions.  We disabled
that diagnostic.  Enable it and add explicit casts.

+++ libipt/src/pt_insn.c
@@ -187,7 +187,8 @@ int pt_insn_next_ip(uint64_t *pip, const struct pt_insn *insn,
 	case ptic_call:
 	case ptic_jump:
 		if (iext->variant.branch.is_direct) {
-			ip += iext->variant.branch.displacement;
+			ip += (uint64_t) (int64_t)
+				iext->variant.branch.displacement;
 			break;
 		}
