+++ libipt/src/pt_query_decoder.c
@@ -653,7 +653,7 @@ int pt_qry_get_offset(const struct pt_query_decoder *decoder, uint64_t *offset)
 	if (!pos)
 		return -pte_nosync;
 
-	*offset = pos - begin;
+	*offset = (uint64_t) (int64_t) (pos - begin);
 	return 0;
 }
 
@@ -671,7 +671,7 @@ int pt_qry_get_sync_offset(const struct pt_query_decoder *decoder,
 	if (!sync)
 		return -pte_nosync;
 
-	*offset = sync - begin;
+	*offset = (uint64_t) (int64_t) (sync - begin);
 	return 0;
 }
